old gen viable/unviable moves take 2, and some other things
The teambuilder will no longer assume moves that are different in old gens were changed by the mod and should therefore default to viable.

You can now add "isViable" to the move data to make sure the move shows as viable. All that happens is it registers as the move having been modified if you add that.

In trying to get old gen mods to automatically show the correct descriptions for old gen moves, I noticed that Dex.mod(gen).moves.get(moveid) does not return the correct old-gen description of a move. However, on main it does (tested via console). This suggests a different issue that will need to be taken care of. However, if that issue gets fixed, the descriptions for old gen mods will work as well, as the changes in js/search.js should implement that.

Fixed another issue in js/search.js where it was calling Dex.getTypeIcon with a null modid. This meant that when different mods had types with the same name, it did not know which mod to use and would search the mod list, defaulting to the first mod in the list that had said type. That issue might be fixed now, though I didn't test it yet.