Team select in teambuilder/challenging now prioritizes that format's mod
While editing the arguments for packedTeamIcons WOULD cause some ripple effects if other files used it (and problems if they're unaware of what mods we are using it for), this same file is the only one to use that method. Now when fetching team icons, it searches for a mod the format may be using.