Fixing "Kommo" and "Ting" aliases
Simply put they cause infinite recursion (Feed Kommo in -> Replace with Kommo-o -> Hyphen present, feed Kommo in) and as such crash builders for mods. This fix checks to see if we're about to feed the same argument(s) into "get" as earlier and if we are, does not perform the internal "get" call.